+++ examples/fork-pi/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @machinen/example-fork-pi
 
+## 0.0.17
+
+### Patch Changes
+
+- Updated dependencies [56ca711]
+- Updated dependencies [8f06495]
+- Updated dependencies [dce77cb]
+  - @machinen/runtime@0.7.1
+
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies [90f20ca]
+- Updated dependencies [90f20ca]
+- Updated dependencies [2ad5f5c]
+  - @machinen/runtime@0.7.0
+
 ## 0.0.15
 
 ### Patch Changes